moved PhoSim astrometry routines into mixins directory
